+++ _tools/travis/autotag.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-set -ex
+set -e
 
 deploykey=~/.ssh/deploy.key
 
@@ -8,9 +8,9 @@ Host github.com
     StrictHostKeyChecking no
     IdentityFile $deploykey
 " >> ~/.ssh/config
-openssl aes-256-cbc -K $encrypted_e129501e13c5_key -iv $encrypted_e129501e13c5_iv -in tool/travis/mackerel-agent.pem.enc -out $deploykey -d
+openssl aes-256-cbc -K $encrypted_e129501e13c5_key -iv $encrypted_e129501e13c5_iv -in _tools/travis/mackerel-agent.pem.enc -out $deploykey -d
 chmod 600 $deploykey
 git config --global user.email "mackerel-developers@hatena.ne.jp"
 git config --global user.name  "mackerel"
 git remote set-url origin git@github.com:mackerelio/mackerel-agent.git
-tool/autotag
+_tools/autotag