+++ checks/checker.go
@@ -46,10 +46,12 @@ type Checker struct {
 
 // Report is what Checker produces by invoking its command.
 type Report struct {
-	Name       string
-	Status     Status
-	Message    string
-	OccurredAt time.Time
+	Name                 string
+	Status               Status
+	Message              string
+	OccurredAt           time.Time
+	NotificationInterval *int32
+	MaxCheckAttempts     *int32
 }
 
 func (c Checker) String() string {
@@ -66,22 +68,26 @@ func (c Checker) Check() (*Report, error) {
 	if stderr != "" {
 		logger.Warningf("Checker %q output stderr: %s", c.Name, stderr)
 	}
-	if err != nil {
-		return nil, err
-	}
 
 	status := StatusUnknown
-	if s, ok := exitCodeToStatus[exitCode]; ok {
-		status = s
+
+	if err != nil {
+		message = err.Error()
+	} else {
+		if s, ok := exitCodeToStatus[exitCode]; ok {
+			status = s
+		}
+
+		logger.Debugf("Checker %q status=%s message=%q", c.Name, status, message)
 	}
 
-	logger.Debugf("Checker %q status=%s message=%q", c.Name, status, message)
-
 	return &Report{
-		Name:       c.Name,
-		Status:     status,
-		Message:    message,
-		OccurredAt: now,
+		Name:                 c.Name,
+		Status:               status,
+		Message:              message,
+		OccurredAt:           now,
+		NotificationInterval: c.Config.NotificationInterval,
+		MaxCheckAttempts:     c.Config.MaxCheckAttempts,
 	}, nil
 }