+++ metrics/windows/interface_test.go
@@ -7,17 +7,17 @@ import (
 )
 
 func TestInterfaceGenerator(t *testing.T) {
-/*
+	/*
 
-	g, _ := NewInterfaceGenerator(5)
+		g, _ := NewInterfaceGenerator(5)
 
-	t.Logf("interval '%s'", g.Interval)
+		t.Logf("interval '%s'", g.Interval)
 
-	_, err := g.Generate()
-	if err != nil {
-		t.Errorf("Generate() failed: %s", err)
-	}
-*/
+		_, err := g.Generate()
+		if err != nil {
+			t.Errorf("Generate() failed: %s", err)
+		}
+	*/
 
 	// TODO
 }