+++ util/filesystem.go
@@ -6,13 +6,15 @@ package util
 
 import (
 	"bufio"
-	"bytes"
+	"os"
 	"os/exec"
 	"regexp"
 	"runtime"
 	"strconv"
 	"strings"
+	"time"
 
+	"github.com/Songmu/timeout"
 	"github.com/mackerelio/mackerel-agent/logging"
 )
 
@@ -58,16 +60,21 @@ func init() {
 // CollectDfValues XXX
 func CollectDfValues(dfColumnSpecs []DfColumnSpec) (map[string]map[string]interface{}, error) {
 	cmd := exec.Command("df", dfOpt...)
-	cmd.Env = append(cmd.Env, "LANG=C")
-
+	cmd.Env = append(os.Environ(), "LANG=C")
+	tio := &timeout.Timeout{
+		Cmd:       cmd,
+		Duration:  15 * time.Second,
+		KillAfter: 5 * time.Second,
+	}
 	// Ignores exit status in case that df returns exit status 1
 	// when the agent does not have permission to access file system info.
-	out, err := cmd.Output()
+	_, stdout, _, err := tio.Run()
+
 	if err != nil {
 		logger.Warningf("'df %s' command exited with a non-zero status: '%s'", strings.Join(dfOpt, " "), err)
 	}
 
-	lineScanner := bufio.NewScanner(bytes.NewReader(out))
+	lineScanner := bufio.NewScanner(strings.NewReader(stdout))
 	filesystems := make(map[string]map[string]interface{})
 
 DF_LINES: